Added a simple parser for parsing commands from the console, including required and optional arguments.
This is a required feature, but will also be handy during development
to for example tell the compiler which module is the main module, and
grant further options to customize the compiler's behaviour.

Ideally this will be replaced at some point with arbitrary compile-time
execution of code, so that one can easily program a custom build system.
For now though, a command line will suffice.