Added language documentation that was previously only on my local machine which is kind of an informal language specification.
Also extended the command line to be able to actually execute
commands now through function pointers. The way they are used means
the compiler should be able to hard-code the calls, while I am able to
reference functions and tie them to commands, which makes them easy to
extend and easy to add new ones.